+++ tests/data/headerFiles/githubDirPath.header.js
@@ -1,0 +1,1 @@
+﻿// %%import<<github::master<<DIR 'twbs/bootstrap/dist/js/'