+++ tests/data/headerFiles/nodeModulesFilePaths.header.js
@@ -1,0 +1,4 @@
+﻿// $import 'randomFile1'
+// $import 'randomFile2'
+// $import 'randomDir1/randomFile1'
+// $import 'randomFile3'